- made isEqualTo comparison functions in amplitude/interactionVertex
  and isobarDecayVertex virtual; comparison operator only defined in
  base class interactionVertex; comparison of mass dependencies in
  isobarDecayVertex does not make sense in current form, because only
  pointer values are compared

- added comparison operator and isEqualTo functions to
  amplitude/particleProperties and particle

+++ generators/genpw.cc
@@ -173,7 +173,7 @@ int main(int argc, char** argv)
   TLorentzVector beam;
   TVector3 vertex;
   double tprime(0.);
-  double qbeam;
+  int qbeam;
   vector<int> q; // array of charges
 
   double Mom=reactConf.lookup("beam.momentum");
